Create a centralised PredicateProducer class
The creation of predicates is done in the
FilterCommandParser class which will get
too long as we add more functionality to filter

Introduce a new PredicateProducer class,
which is suitably added in the Predicate package class.
This class centralises the creation of predicates
based on its respective prefixes
(i.e. PREFIX_NAME, PREFIX_REMARK)

Centralising predicate creation enhances
code readability, maintainability, and scalability.